+++ java-server/src/main/java/com/cooksys/ftd/week3/command/RegisterUser.java
@@ -9,6 +9,10 @@ import org.eclipse.jdt.annotation.NonNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.cooksys.ftd.week3.db.DBConnection;
+import com.cooksys.ftd.week3.db.dao.FileDao;
+import com.cooksys.ftd.week3.db.dao.UserDao;
+import com.cooksys.ftd.week3.db.model.User;
 import com.cooksys.ftd.week3.transactions.ServerMessage;
 
 @XmlRootElement
@@ -19,14 +23,27 @@ public class RegisterUser implements AbstractCommand {
 	private PrintWriter writer;
 	private Map<String, Object> args;
 	
+	private UserDao userDao = new UserDao();
+	private FileDao fileDao = new FileDao();
+	
 	public RegisterUser(){
 		super();
 	}
 	
 	@Override
 	public ServerMessage executeCommand(@NonNull Map<String, Object> args) {
-		Server
-		return null;
+		User user = new User();
+		user.setUsername((String)args.get(User.USERNAME_COLUMN));
+		user.setPassword((String)args.get(User.PASSWORD_COLUMN));
+		
+		boolean inserted = userDao.insertUser(user, DBConnection.connection);
+		
+		ServerMessage sm = new ServerMessage();
+		sm.setError(!inserted);
+		if(!inserted) {
+			sm.setMessage("Error inserting user credentials into database");
+		}
+		return sm;
 	}
 
 	@Override
@@ -47,5 +64,21 @@ public class RegisterUser implements AbstractCommand {
 	@Override
 	public Map<String, Object> getArgs() {
 		return this.args;
+	}
+
+	public FileDao getFileDao() {
+		return fileDao;
+	}
+
+	public void setFileDao(FileDao fileDao) {
+		this.fileDao = fileDao;
+	}
+
+	public UserDao getUserDao() {
+		return userDao;
+	}
+
+	public void setUserDao(UserDao userDao) {
+		this.userDao = userDao;
 	}
 }
